Added colour picker

+++ src/main/java/com/HomeAuto/backend/LimitlessLED.java
@@ -9,24 +9,24 @@ public class LimitlessLED {
 
         if (command.equals("on")) {
             if (group == 1){
-                UDPTransmit.sendData[0] = 0x45;
+                UDPTransmit.sendData[0] = (byte)0x45;
             }
             if (group == 2){
-                UDPTransmit.sendData[0] = 0x47;
+                UDPTransmit.sendData[0] = (byte)0x47;
             }
             else{
-                UDPTransmit.sendData[0] = 0x42;
+                UDPTransmit.sendData[0] = (byte)0x42;
             }
         }
         if (command.equals("off")){
             if (group == 1){
-                UDPTransmit.sendData[0] = 0x46;
+                UDPTransmit.sendData[0] = (byte)0x46;
             }
             if (group == 2){
-                UDPTransmit.sendData[0] = 0x48;
+                UDPTransmit.sendData[0] = (byte)0x48;
             }
             else{
-                UDPTransmit.sendData[0] = 0x41;
+                UDPTransmit.sendData[0] = (byte)0x41;
             }
         }
         if (command.equals("white")){
@@ -35,13 +35,13 @@ public class LimitlessLED {
         }
         if (command.equals("green")){
             lightControl(group,"on");
-            UDPTransmit.sendData[0] = 0x40;
-            UDPTransmit.sendData[1] = 0x60;
+            UDPTransmit.sendData[0] = (byte)0x40;
+            UDPTransmit.sendData[1] = (byte)0x60;
         }
         if (command.equals("blue")){
             lightControl(group,"on");
-            UDPTransmit.sendData[0] = 0x40;
-            UDPTransmit.sendData[1] = 0x20;
+            UDPTransmit.sendData[0] = (byte)0x40;
+            UDPTransmit.sendData[1] = (byte)0x20;
         }
 
         UDPTransmit.udpSend(host, port);
